[C++] Change Vector3D fields from double to float, to have alignment eq. 4 (to support MSVC 2010), plus minor review fix

+++ tests/native_type_test_impl.h
@@ -3,12 +3,12 @@
 
 namespace Native {
   struct Vector3D {
-    double x;
-    double y;
-    double z;
+    float x;
+    float y;
+    float z;
 
     Vector3D() { x = 0; y = 0; z = 0; };
-    Vector3D(double x, double y, double z) { this->x = x; this->y = y; this->z = z; }
+    Vector3D(float x, float y, float z) { this->x = x; this->y = y; this->z = z; }
   };
 }
